+++ packages/amnis-process/src/auth/auth.login.test.ts
@@ -1,8 +1,8 @@
-import { dbmemory } from '@amnis/db';
-import { fsmemory } from '@amnis/fs';
 import {
   AuthLogin,
+  databaseMemory,
   dataInitial,
+  filesystemMemory,
   IoInput,
   ioProcess,
   Profile,
@@ -21,8 +21,8 @@ const io = ioProcess(
   {
     store: storeSetup(),
     validators: validateSetup([schemaAuth]),
-    database: dbmemory,
-    filesystem: fsmemory,
+    database: databaseMemory,
+    filesystem: filesystemMemory,
     crypto: cryptoWeb,
   },
   {
@@ -31,7 +31,7 @@ const io = ioProcess(
 );
 
 beforeAll(async () => {
-  await dbmemory.create(dataInitial());
+  await databaseMemory.create(dataInitial());
 });
 
 test('should login as administrator', async () => {